+++ kittens/panel/main.py
@@ -3,11 +3,13 @@
 
 import sys
 from collections.abc import Callable
-from typing import Any
+from contextlib import suppress
+from functools import partial
+from typing import Any, Mapping, Sequence
 
 from kitty.cli import parse_args
 from kitty.cli_stub import PanelCLIOptions
-from kitty.constants import appname, is_macos, is_wayland
+from kitty.constants import appname, is_macos, is_wayland, kitten_exe
 from kitty.fast_data_types import (
     GLFW_EDGE_BOTTOM,
     GLFW_EDGE_CENTER,
@@ -27,46 +29,46 @@ from kitty.fast_data_types import (
 )
 from kitty.os_window_size import WindowSizeData, edge_spacing
 from kitty.types import LayerShellConfig
-from kitty.typing import EdgeLiteral
+from kitty.typing import BossType, EdgeLiteral
 
 OPTIONS = r'''
 --lines
-type=int
 default=1
 The number of lines shown in the panel. Ignored for background, centered, and vertical panels.
+If it has the suffix :code:`px` then it sets the height of the panel in pixels instead of lines.
 
 
 --columns
-type=int
 default=1
 The number of columns shown in the panel. Ignored for background, centered, and horizontal panels.
+If it has the suffix :code:`px` then it sets the width of the panel in pixels instead of columns.
 
 
 --margin-top
 type=int
 default=0
-Request a given top margin to the compositor.
+Set the top margin for the panel, in pixels. Has no effect for bottom edge panels.
 Only works on a Wayland compositor that supports the wlr layer shell protocol.
 
 
 --margin-left
 type=int
 default=0
-Request a given left margin to the compositor.
+Set the left margin for the panel, in pixels. Has no effect for right edge panels.
 Only works on a Wayland compositor that supports the wlr layer shell protocol.
 
 
 --margin-bottom
 type=int
 default=0
-Request a given bottom margin to the compositor.
+Set the bottom margin for the panel, in pixels. Has no effect for top edge panels.
 Only works on a Wayland compositor that supports the wlr layer shell protocol.
 
 
 --margin-right
 type=int
 default=0
-Request a given right margin to the compositor.
+Set the right margin for the panel, in pixels. Has no effect for left edge panels.
 Only works on a Wayland compositor that supports the wlr layer shell protocol.
 
 
@@ -111,7 +113,7 @@ you to specify which output is used, by name. If not specified the compositor wi
 output automatically, typically the last output the user interacted with or the primary monitor.
 
 
---class
+--class --app-id
 dest=cls
 default={appname}-panel
 condition=not is_macos
@@ -144,6 +146,19 @@ If :option:`--edge` is set to :code:`background`, this option has no effect.
 type=bool-set
 On a Wayland compositor that supports the wlr layer shell protocol, override the default exclusive zone.
 This has effect only if :option:`--edge` is set to :code:`top`, :code:`left`, :code:`bottom` or :code:`right`.
+
+
+--single-instance -1
+type=bool-set
+If specified only a single instance of the panel will run. New
+invocations will instead create a new top-level window in the existing
+panel instance.
+
+
+--instance-group
+Used in combination with the :option:`--single-instance` option. All
+panel invocations with the same :option:`--instance-group` will result
+in new panels being created in the first panel instance within that group.
 
 
 --debug-rendering
@@ -214,25 +229,39 @@ def initial_window_size_func(opts: WindowSizeData, cached_values: dict[str, Any]
             xscale = yscale = 1
         global window_width, window_height
         monitor_width, monitor_height = glfw_primary_monitor_size()
+        x = dual_distance(args.columns, min_cell_value_if_no_pixels=1)
+        rwidth = x[1] if x[1] else (x[0] * cell_width / xscale)
+        x = dual_distance(args.lines, min_cell_value_if_no_pixels=1)
+        rheight = x[1] if x[1] else (x[0] * cell_width / yscale)
 
         if args.edge in {'left', 'right'}:
             spacing = es('left') + es('right')
-            window_width = int(cell_width * args.columns / xscale + (dpi_x / 72) * spacing + 1)
+            window_width = int(rwidth + (dpi_x / 72) * spacing + 1)
             window_height = monitor_height
         elif args.edge in {'top', 'bottom'}:
             spacing = es('top') + es('bottom')
-            window_height = int(cell_height * args.lines / yscale + (dpi_y / 72) * spacing + 1)
+            window_height = int(rheight + (dpi_y / 72) * spacing + 1)
             window_width = monitor_width
         elif args.edge in {'background', 'center'}:
             window_width, window_height = monitor_width, monitor_height
         else:
             x_spacing = es('left') + es('right')
-            window_width = int(cell_width * args.columns / xscale + (dpi_x / 72) * x_spacing + 1)
+            window_width = int(rwidth + (dpi_x / 72) * x_spacing + 1)
             y_spacing = es('top') + es('bottom')
-            window_height = int(cell_height * args.lines / yscale + (dpi_y / 72) * y_spacing + 1)
+            window_height = int(rheight + (dpi_y / 72) * y_spacing + 1)
         return window_width, window_height
 
     return initial_window_size
+
+
+def dual_distance(spec: str, min_cell_value_if_no_pixels: int = 0) -> tuple[int, int]:
+    with suppress(Exception):
+        return int(spec), 0
+    if spec.endswith('px'):
+        return min_cell_value_if_no_pixels, int(spec[:-2])
+    if spec.endswith('c'):
+        return int(spec[:-1]), 0
+    return min_cell_value_if_no_pixels, 0
 
 
 def layer_shell_config(opts: PanelCLIOptions) -> LayerShellConfig:
@@ -247,10 +276,11 @@ def layer_shell_config(opts: PanelCLIOptions) -> LayerShellConfig:
     focus_policy = {
         'not-allowed': GLFW_FOCUS_NOT_ALLOWED, 'exclusive': GLFW_FOCUS_EXCLUSIVE, 'on-demand': GLFW_FOCUS_ON_DEMAND
     }.get(opts.focus_policy, GLFW_FOCUS_NOT_ALLOWED)
+    x, y = dual_distance(opts.columns, min_cell_value_if_no_pixels=1), dual_distance(opts.lines, min_cell_value_if_no_pixels=1)
     return LayerShellConfig(type=ltype,
                             edge=edge,
-                            x_size_in_cells=max(1, opts.columns),
-                            y_size_in_cells=max(1, opts.lines),
+                            x_size_in_cells=x[0], x_size_in_pixels=x[1],
+                            y_size_in_cells=y[0], y_size_in_pixels=y[1],
                             requested_top_margin=max(0, opts.margin_top),
                             requested_left_margin=max(0, opts.margin_left),
                             requested_bottom_margin=max(0, opts.margin_bottom),
@@ -259,6 +289,18 @@ def layer_shell_config(opts: PanelCLIOptions) -> LayerShellConfig:
                             requested_exclusive_zone=opts.exclusive_zone,
                             override_exclusive_zone=opts.override_exclusive_zone,
                             output_name=opts.output_name or '')
+
+
+def handle_single_instance_command(boss: BossType, sys_args: Sequence[str], environ: Mapping[str, str], notify_on_os_window_death: str | None = '') -> None:
+    from kitty.tabs import SpecialWindow
+    args, items = parse_panel_args(list(sys_args[1:]))
+    items = items or [kitten_exe(), 'run-shell']
+    lsc = layer_shell_config(args)
+    os_window_id = boss.add_os_panel(lsc, args.cls, args.name)
+    if notify_on_os_window_death:
+        boss.os_window_death_actions[os_window_id] = partial(boss.notify_on_os_window_death, notify_on_os_window_death)
+    tm = boss.os_window_map[os_window_id]
+    tm.new_tab(SpecialWindow(cmd=items, env=dict(environ)))
 
 
 def main(sys_args: list[str]) -> None:
@@ -279,6 +321,8 @@ def main(sys_args: list[str]) -> None:
     for override in args.override:
         sys.argv.extend(('--override', override))
     sys.argv.append('--override=linux_display_server=auto')
+    if args.single_instance:
+        sys.argv.append('--single-instance')
     sys.argv.extend(items)
     from kitty.main import main as real_main
     from kitty.main import run_app