+++ tools/utils/tar.go
@@ -10,6 +10,7 @@ import (
 	"io/fs"
 	"os"
 	"path/filepath"
+	"runtime"
 	"strings"
 )
 
@@ -19,15 +20,166 @@ type TarExtractOptions struct {
 	DontPreservePermissions bool
 }
 
+func volnamelen(path string) int {
+	return len(filepath.VolumeName(path))
+}
+
+func EvalSymlinksThatExist(path string) (string, error) {
+	volLen := volnamelen(path)
+	pathSeparator := string(os.PathSeparator)
+
+	if volLen < len(path) && os.IsPathSeparator(path[volLen]) {
+		volLen++
+	}
+	vol := path[:volLen]
+	dest := vol
+	linksWalked := 0
+	for start, end := volLen, volLen; start < len(path); start = end {
+		for start < len(path) && os.IsPathSeparator(path[start]) {
+			start++
+		}
+		end = start
+		for end < len(path) && !os.IsPathSeparator(path[end]) {
+			end++
+		}
+
+		// On Windows, "." can be a symlink.
+		// We look it up, and use the value if it is absolute.
+		// If not, we just return ".".
+		isWindowsDot := runtime.GOOS == "windows" && path[volnamelen(path):] == "."
+
+		// The next path component is in path[start:end].
+		if end == start {
+			// No more path components.
+			break
+		} else if path[start:end] == "." && !isWindowsDot {
+			// Ignore path component ".".
+			continue
+		} else if path[start:end] == ".." {
+			// Back up to previous component if possible.
+			// Note that volLen includes any leading slash.
+
+			// Set r to the index of the last slash in dest,
+			// after the volume.
+			var r int
+			for r = len(dest) - 1; r >= volLen; r-- {
+				if os.IsPathSeparator(dest[r]) {
+					break
+				}
+			}
+			if r < volLen || dest[r+1:] == ".." {
+				// Either path has no slashes
+				// (it's empty or just "C:")
+				// or it ends in a ".." we had to keep.
+				// Either way, keep this "..".
+				if len(dest) > volLen {
+					dest += pathSeparator
+				}
+				dest += ".."
+			} else {
+				// Discard everything since the last slash.
+				dest = dest[:r]
+			}
+			continue
+		}
+
+		// Ordinary path component. Add it to result.
+
+		if len(dest) > volnamelen(dest) && !os.IsPathSeparator(dest[len(dest)-1]) {
+			dest += pathSeparator
+		}
+
+		dest += path[start:end]
+
+		// Resolve symlink.
+
+		fi, err := os.Lstat(dest)
+		if err != nil {
+			if os.IsNotExist(err) {
+				if end < len(path) {
+					dest += path[end:]
+				}
+				return filepath.Clean(dest), nil
+			}
+			return "", err
+		}
+
+		if fi.Mode()&fs.ModeSymlink == 0 {
+			if !fi.Mode().IsDir() && end < len(path) {
+				return "", fmt.Errorf("%s is not a directory while resolving symlinks in %s", dest, path)
+			}
+			continue
+		}
+
+		// Found symlink.
+
+		linksWalked++
+		if linksWalked > 255 {
+			return "", fmt.Errorf("EvalSymlinksThatExist: too many symlinks in %s", path)
+		}
+
+		link, err := os.Readlink(dest)
+		if err != nil {
+			return "", err
+		}
+
+		if isWindowsDot && !filepath.IsAbs(link) {
+			// On Windows, if "." is a relative symlink,
+			// just return ".".
+			break
+		}
+
+		path = link + path[end:]
+
+		v := volnamelen(link)
+		if v > 0 {
+			// Symlink to drive name is an absolute path.
+			if v < len(link) && os.IsPathSeparator(link[v]) {
+				v++
+			}
+			vol = link[:v]
+			dest = vol
+			end = len(vol)
+		} else if len(link) > 0 && os.IsPathSeparator(link[0]) {
+			// Symlink to absolute path.
+			dest = link[:1]
+			end = 1
+			vol = link[:1]
+			volLen = 1
+		} else {
+			// Symlink to relative path; replace last
+			// path component in dest.
+			var r int
+			for r = len(dest) - 1; r >= volLen; r-- {
+				if os.IsPathSeparator(dest[r]) {
+					break
+				}
+			}
+			if r < volLen {
+				dest = vol
+			} else {
+				dest = dest[:r]
+			}
+			end = 0
+		}
+	}
+	return filepath.Clean(dest), nil
+}
+
 func ExtractAllFromTar(tr *tar.Reader, dest_path string, optss ...TarExtractOptions) (count int, err error) {
 	opts := TarExtractOptions{}
 	if len(optss) > 0 {
 		opts = optss[0]
 	}
-	dest_path, err = filepath.Abs(dest_path)
-	if err != nil {
+	if !filepath.IsAbs(dest_path) {
+		if dest_path, err = filepath.Abs(dest_path); err != nil {
+			return
+		}
+	}
+	if dest_path, err = filepath.EvalSymlinks(dest_path); err != nil {
 		return
 	}
+	dest_path = filepath.Clean(dest_path)
 
 	mode := func(hdr int64) fs.FileMode {
 		return fs.FileMode(hdr) & (fs.ModePerm | fs.ModeSetgid | fs.ModeSetuid | fs.ModeSticky)
@@ -43,6 +195,7 @@ func ExtractAllFromTar(tr *tar.Reader, dest_path string, optss ...TarExtractOpti
 		count++
 		return
 	}
+	needed_prefix := dest_path + string(os.PathSeparator)
 
 	for {
 		var hdr *tar.Header
@@ -55,19 +208,13 @@ func ExtractAllFromTar(tr *tar.Reader, dest_path string, optss ...TarExtractOpti
 			return count, err
 		}
 		dest := hdr.Name
-		dest = strings.TrimLeft(dest, "/")
-		if !filepath.IsLocal(dest) {
-			continue
+		if !filepath.IsAbs(dest) {
+			dest = filepath.Join(dest_path, dest)
 		}
-		dest = filepath.Join(dest_path, dest)
-		if dest, err = filepath.EvalSymlinks(dest); err != nil {
-			if os.IsNotExist(err) {
-				err = nil
-			} else {
-				return count, err
-			}
+		if dest, err = EvalSymlinksThatExist(dest); err != nil {
+			return count, err
 		}
-		if !strings.HasPrefix(filepath.Clean(dest), filepath.Clean(dest_path)+string(os.PathSeparator)) {
+		if !strings.HasPrefix(dest, needed_prefix) {
 			continue
 		}
 		switch hdr.Typeflag {